Add --json-pointer option
DongFangHong's script puts the reddit API response inside a property named
"json" in the top-level object on each line. This is fine, but it's not great
if people want to use this importer directly on the API response.

The new --json-pointer option locates the reddit API response within the
top-level object. That way by default it can use the root, and DongFangHong's
script becomes a special case handled with --json-pointer=/json